Create Scott Lambert
Welcome to the Computing Core Skills and Open Source Software part of CPU4000

In this part of the module we'll familiarise you with some of the tools used by the software industry, and develop your skills and understanding of some industry practices.

In this first session we'll be setting up a blog using Wordpress, and writing a few posts.

It may surprise you that in every job in IT you actually spend a lot more time writing in English than you ever do in a programming language - so written communication is a core skill. The best way to improve it is to practice it, and so every week in class will involve some writing. Each of these written pieces will contribute to your final assessment.

As this is the first class, you haven't had any set reading (or viewing), but note in every future class there will always be videos to watch or articles to read BEFORE coming to class.

Its important you do the reading or watching as we won't be going over the material again in class.

Wordpress offers free blog hosting, and so is a good place to create your blog for this class.

Everything you need to get started is at https://wordpress.com/

Once you've created a site, we need to know where to find it so we can read it!

To do this you need to fill in this form: oss.ly/wpform

Computing is a really diverse subject, and touches every aspect of the way we live.

In class, we'll brainstorm ideas for interesting topics.

Afterwards, write a post about one topic that interests you.

Make sure to:

Write using full sentences
Find some good articles to link to from your post
Explain why it interests you in particular
Consider the implications - what will happen in the future?

There are hundreds of different kinds of roles and jobs in computing, or related to computing, as well as quite a few different ways to be self-employed, or engage as a volunteer.

In class we'll brainstorm as many options as we can think of.

Afterwards, pick an option that looks interesting, explore it and write up your findings as a blog post.

Remember:

Write in full sentences
Find good articles to link to from within your post
Consider both the upsides and downsides
Ask whether this is something you could imagine yourself doing
Think about the kind of qualities that the role would require